tests: move up mockFilesystem so it get be re-used in other tests
Note: removed `$this->mockOutput = '';` as it was a no-op, that property
doesn't exist (it was probably a typo, but I guess we don't need it
anyway)
