+++ tests/TestCase.php
@@ -5,6 +5,8 @@ declare(strict_types=1);
 namespace Barryvdh\LaravelIdeHelper\Tests;
 
 use Illuminate\Console\Command;
+use Illuminate\Filesystem\Filesystem;
+use Mockery;
 use Orchestra\Testbench\TestCase as BaseTestCase;
 use Spatie\Snapshots\MatchesSnapshots;
 use Symfony\Component\Console\Tester\CommandTester;
@@ -12,6 +14,8 @@ use Symfony\Component\Console\Tester\CommandTester;
 abstract class TestCase extends BaseTestCase
 {
     use MatchesSnapshots;
+
+    protected $mockFilesystemOutput;
 
     /**
      * The `CommandTester` is directly returned, use methods like
@@ -49,5 +53,31 @@ abstract class TestCase extends BaseTestCase
     protected function assertMatchesTxtSnapshot(?string $actualContent)
     {
         $this->assertMatchesSnapshot($actualContent, new SnapshotTxtDriver());
+    }
+
+    protected function mockFilesystem()
+    {
+        $mockFilesystem = Mockery::mock(Filesystem::class);
+
+        $mockFilesystem
+            ->shouldReceive('get')
+            ->andReturnUsing(function ($file) {
+                return file_get_contents($file);
+            });
+
+        $mockFilesystem
+            ->shouldReceive('put')
+            ->with(
+                Mockery::any(),
+                Mockery::any()
+            )
+            ->andReturnUsing(function ($path, $contents) {
+                $this->mockFilesystemOutput .= $contents;
+
+                return strlen($contents);
+            });
+
+        $this->instance(Filesystem::class, $mockFilesystem);
+        $this->instance('files', $mockFilesystem);
     }
 }
